[BE] feat: jpql 변경 완료 NoticeRepository의 native sql을 jpql로 변경 완료했습니다.

+++ BE/src/main/java/com/softeer/BE/controller/NoticeController.java
@@ -14,7 +14,7 @@ import java.util.List;
 public class NoticeController {
 
     private final NoticeService noticeService;
-    private static final Long DEFAULT_PAGE_SIZE = 10L;
+    private static final Integer DEFAULT_PAGE_SIZE = 10;
 
     @GetMapping("/{noticeId}")
     public ApiResponse<NoticeDto> getNotice(@PathVariable Long noticeId) {
@@ -24,8 +24,8 @@ public class NoticeController {
 
     // cursorId는 클라이언트가 받은 Notice의 가장 낮은 id를 의미하고, 처음 요청시 0을 보냄
     @GetMapping("/list")
-    public ApiResponse<List<NoticeDto>> getNoticeList(@RequestParam Long cursorId, @RequestParam Long pageSize) {
-        if (pageSize <= 0L) pageSize = DEFAULT_PAGE_SIZE;
+    public ApiResponse<List<NoticeDto>> getNoticeList(@RequestParam Long cursorId, @RequestParam Integer pageSize) {
+        if (pageSize <= 0) pageSize = DEFAULT_PAGE_SIZE;
         List<NoticeDto> noticeDtoList = noticeService.getNoticeList(cursorId, pageSize);
         return ApiResponse.onSuccess(noticeDtoList);
     }
